requirements.txt
streamlit
pandas
numpy
Pillow
opencv-python
scikit-learn
matplotlib